requirements.txt
torch==2.4.1
transformers==4.45.2
nltk==3.9.1
flask==3.0.3
werkzeug==3.0.3
PyPDF2==3.0.1
python-docx==1.1.2
markupsafe==2.1.5
absl-py==2.1.0
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
blis==1.0.1
catalogue==2.0.10
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.19.0
colorama==0.4.6
comtypes==1.4.5
confection==0.1.5
constantly==23.10.4
contourpy==1.3.0
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==5.1
django-anymail==11.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dominate==2.9.1
et-xmlfile==1.1.0
filelock==3.15.4
flake8==7.1.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
google-tts==1.0.1
greenlet==3.0.3
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.25.1
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
kombu==5.4.0
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
limits==3.13.0
llvmlite==0.43.0
lxml==5.2.2
Mako==1.3.5
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
mitreapi==1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
mysqlclient==2.2.4
namex==0.0.8
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parsel==1.9.1
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.47
Protego==0.3.1
protobuf==4.25.5
psutil==6.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.1.0
pyparsing==3.1.4
PyPDF2==3.0.1
pypiwin32==223
PySocks==1.7.1
pystray==0.19.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
pyttsx3==2.90
pytz==2024.2
pywifi==1.1.12
pywin32==306
PyYAML==6.0.2
queuelib==1.7.0
redis==5.0.8
regex==2024.9.11
reportlab==4.2.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
rich==13.9.2
sacremoses==0.1.1
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.0
Scrapy==2.11.2
selenium==4.25.0
sentencepiece==0.2.0
service-identity==24.1.0
setuptools==71.0.4
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.3.7
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
sqlparse==0.5.1
srsly==2.4.8
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.20.0
torch==2.4.1
torchtext==0.18.0
tqdm==4.66.5
transformers==4.45.2
trio==0.26.2
trio-websocket==0.11.1
tweepy==4.14.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
visitor==0.1.3
w3lib==2.2.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
zipp==3.20.2
zope.interface==6.4.post2
